wip: github & jira integration sync in DB
